qps: fix bogus unlinking of files
small allocation maps that were read-only should not be unlinked from the
disk because they belong to a snapshot. Meaning that:

* qps_map_recycle must NOT EVER unlink files. If we do that, when a GC
  moves all allocation away from an old .qps file, we may remove it too
  early.

* When recycling an ID, we cannot reuse it instantaneously, if we do so,
  the .qps file on disk will be destroyed on reuse, possibly overwriting a
  file we didn't want to lose.

That's why, when the map we recycle is a small allocation map, we put it
in a list of quarantined ids, that is unquarantined after each snapshot is
done so that the ids can be reused.

This isn't required for paged maps because their storage is a .qpz.

Also be on the safe side when unlinking a file in qps_map_m_create_raw, it
cannot hurt: IOW if unlinkat fails for another reason than ENOENT, bail
out.

Closes: #7900
Change-Id: I41e024d2f5efd19878301b09b0f29d5c309f697a
Signed-off-by: Pierre Habouzit <pierre.habouzit@intersec.com>