ichttp: harmonize ichttp_register with ic_register macro
Harmony between ic and ichttp macros is a logical expectation, and having a
macro to use a generic name for the implementation of the registered RPC is
a basic need, otherwise it would not be defined separately in each products,
so this change will help to factorize some code.

Note:
~~~~~
I expect the lib-common update into mmsx:master not to be too much painful
and will perform it as soon this patch is merged.

Change-Id: Iee6c0ea725e493bec28f3ddf42063913b34aad60
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>