qps: fix issue with mapno recycling.
When a mapno was recycled without quarantine, it was added neither in
qps->no_free (free list of mapno) nor in qps->no_blocked (quarantine list
of mapno that will be merged in the free list after next snapshot),
therefore it was leaked (until the next startup of the qps)

As a consequence, the free list was often wrongly empty, leading to the
allocation of a new mapno and the reallocation of the page header
structure. Since that page header structure is malloced using p_new(),
its size is limited to 1GB (which correspond to 1365 maps).

At startup the freelist is properly populated thus the issue is always
solved by a restart of the qps, but since affected spool still have
abnormally high values of mapnos, an allocation of max(mapno) * 12 * 65536
is always done at startup. The issue mostly affect affects instances with
a long runtime or with frequent snapshots.

This patch now does the following:
- if quarantine is requested, add the mapno in no_blocked
- if quarantine is not requested, add the mapno in no_free (with respect
  to the fact no_free is sorted).

That way, mapnos are always reused and their value shouldn't diverge
anymore.

Change-Id: I1b2a5ee48026b0787d600364c3b1163aea6e168e
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>