membitcount: get rid of ifunc usage.
ifunc does not works with valgrind 3.7 and that's really annoying. Instead
of using ifunc, use an global variable to store the pointer to membitcount
and resolve the actual function by hand.

The new implementation still use the 'target' attribute of gcc to generate
optimal code for certain targets and thus requires at least gcc 4.4 in
order to provide an optimal version of membitcount for the machine on
which the code run.

Note that the ifunc solution required gcc 4.6 that is not supported by
redhat (and will not be anytime soon), but the new solution is optimal
with gcc 4.4 which is the supported compiler on redhat 6.

Change-Id: Ibf1a3cbb2911a824c985f5f60332f52993b18806
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>