ichannel: Add a soft mode in ic_watch_activity.
This new mode allows to be notified in the on_event callback of an
ichannel with IC_EVT_NOACT when an inactivity period of timeout_soft
milliseconds is detected, and with IC_EVT_ACT when an activity occurs
after a period of inactivity.

See ic_watch_activity documentation for more details.

Change-Id: I3d5d1a91c2fbcf49e9e17f1c001e684d2270af41
Refs: #7521