arith: import SSE scanners from qhats.
For most function, we only keep the SSE2 version because those versions
are experimentally faster than the SSE4.1 one. Only the count_non_zero64
have a SSE4.1 version because it has no SSE2 version. We load that version
using the same trick as used for membitcount: a pseudo ifunc.

We don't keep the non-SSE version either: all our builds are done with
SSE2 activated.

Change-Id: I23bc2c617c33ffc2e258533f3aedcf6977f5eab1
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>