http: Mishandling of missing data in http requests
Problem:
~~~~~~~
if a HTTP frame contains extra CRLF, our parsing is too strict
and the consequence is we are expecting other request.
In response we send a concatenation of HTTP 200 OK and HTTP 400
Bad request.

Solution:
~~~~~~~~
Ignore extra CRLF as it is advised in RFC 2616 chapter 4.1 to avoid problems
with customer buggy frames.

Refs: #9228
Change-Id: I921018c83616c059e95fc8c34a3ef6a534fd53ce