qkv: wah: fix bad check in wah_flatten_last_run.
Reminder:
---------
WAH is a sequence of runs. A run is at least 2 words longs with the
following format:
- first word:
    represent a large sequence of the same bit. It has the following
    fields:
      * bit value: the repeated bit
      * words: the number of consecutive words with uniformly that bit
- second word:
    number of literal (uncompressed) words that follow. This word is
    followed by the advertised number of words

Implementation:
---------------
The very first run is stored inlined in the wah_t structure excepted its
literal words (if any). The following runs are stored in an array of
words.

We keep the position of the last run in the wah_t structure. That position
is -1 if the last run is the first (inlined) one.

Bug:
----
The bug was caused by the fact wah_flatten_last_run supposed the last run
was always at a strictly positive offset in the word vector. This is buggy
because the first run may have no literal words and thus if there's only 2
runs in the WAH, the second (and last one) starts at position 0.

Also add a non-regression test.

Closes: #9437
Change-Id: I239bd25f414fc7562272c11ae5ce3cbb3aa14871
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>