bithacks: fix bsr().
We didn't compute the offset properly when the bit was in the first
scanned word. We have:
 * Len: the size of data remaining to scan
 * Up: the most significant bit to scan in the current word
 * Low: the least significant bit to scan
 * Pos: the position of the matching bit

The actual offset of the bit in the array is its Pos relative to the first
bit to scan (Pos - Low) plus the offset of Low in the memory to scan. The
offset of Low is the size of remaining memory after consuming what is
between Low and Up, thus Len - (Up - Low).

So, the actual offset of Pos in the memory is:
 (Pos - Low) + (Len - (Up - Low)) = Len + Pos - Up;

Change-Id: I190431c2df794a27983c4ea7eeb0397d717a0a74
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>