bit-buf: fix broken bb_init_sb API.
This API led to a dangling pointer in the sb since both the sb kept
pointer to the data. In case of a realloc by one of them, the other would
have ended with a pointer to the memory before reallocation which is
dangerous.

So, bb_init_sb performs an ownership transfer by reseting the sb and a new
function bb_transfer_to_sb does the exact reverse by transferring the
memory to the sb and resetting the bb.

Also fix the only call site and add a missing sb_init() in bb_init().

Change-Id: I45727998ccf87b77c75a2b9bfda475118083e234
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>