bit-buf: rework the implementation to be 64bits-based.
The new implementation is not based on sb_t anymore, it uses its own
memory allocation. Instead of being byte based, it is now 64bits based,
leading to much less arithmetic for simplement operations such as bit
addition (well, actually, this is not true for the current big-endian
based implementation).

The support for marking certain offsets has been dropped in the rework,
but function stub are still here in order to restore that support sometime
in the future.

I just hope this didn't break anything is asn1-per (tests pass).

Change-Id: Iaaed744cb258b4b8b6e34438445ac9b4bd9484ad
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>