qkv: wah: fix leak when copying a wah.
This only occurs when evaluating .or() expressions in QKV.

Since we wiped the vector 'data' (by copying the vector from the source
and then initializing it) we leaked the pointer of that vector. The new
implementation remembers the previous vector and restore it, allowing
memory reuse (and thus avoiding allocations). However, the consequence of
that patch is that wah_copy() requires an initialized map as destination.

Closes: #9541
Change-Id: Ib80f6d09d32a4322e46093c4a21afde87efe8840
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>