qps: don't remove immediately recycled files.
Some files may be emptied concurrently to the snapshot. As a consequence
they were marked for recycling at the end of the snapshot and immediately
deleted by the cleanup.

This is buggy because the files are still referenced by the current (just
finished) snapshot. As a consequence the snapshot cannot be opened
anymore.

Fix this by performing the cleanup before traversing the maps looking for
empty ones. That way, we're sure we're only removing maps that were marked
for recycling before the beginning of the snapshot.

Change-Id: Iae991e06dc7820b14debbcd46d5687354cb86f7f
Signed-off-by: Sebastien Boulle <sebastien.boulle@intersec.com>