qps: put empty maps in quarantine at the end of the snapshot.
If not done, the maps can still be deleted when closing since the id is
put in the free-list of ids. Thus the map id can be reused and in that
case the new map will not be read-only and will be removed when closing
the qps.

Correct behavior is to "block" the id by putting it in the quarantine list
of map ids that are "free" but still referenced by current snapshot, that
way the map will be effectively freed during the next snapshot without
risking early deletion.

Change-Id: I3abab0d1b83a7de9317ce696a1b4cbcb01e9b3b4
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>