qps: makes map checking stricter.
Checking was done within internal operations meaning that could be done
between the memory mapping change and the handle allocation change,
leading to partial results which meant the check was often meaningless.

The checks has been moved at the beginning and end of the public API
implementations and thus are always done in situations where the QPS is in
a consistent state.

As a consequence we can also get rid of the 'report_ro_errs' parameter
which only meant "be strict", and we can use real error reporting (well,
panic here) instead of an uncatchable trace.

Change-Id: I05c8e983aff6e5b9b51c0f4e9e0e6c44fb2b2959
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>