qps: fix __qps_check_consistency having side effect on the qps.
When loading the meta.qps file we did write restore the remaining count
from the meta.qps to the map which is shared mapped memory. As a
consequence, __qps_check_consistency() set back the remaining count of all
maps for all processes (probably a single one) that opened the qps to the
value it has the last time the QPS was snapshotted.

In case __qps_check_consistency() is not executed concurrently to a
snapshot, the consequence was a corruption of runtime data, the QPS
believing it had more data remaining than actually remaining. That
corruption is then persisted at the next snapshot. Affected maps may then
never be deleted by the garbage collector since their remaining count
cannot read 0 (actually remaining data is smaller or equal to the
remaining count kept by the QPS).

In case __qps_check_consistency() is execute concurrently to a snapshot,
this may have the consequences described above or some more complex
implications if some maps get deleted then recreated (but the probably is
very very very very small, so I don't think this can happen at all).

The fix create a notion of "read-only" loading used checking consistency.
As a consequence the map checking code has been a bit modified and somehow
integrated to the qps_load_meta function in order to be able to check
read-only maps. We do also force a full check of the maps when loading a
QPS in !NDEBUG mode or when the map is read-only so that we can have some
early reporting of corrupted QPS.

Change-Id: Id89868f91699bd5896705e1a946d6aecb21c8057
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>