qps: fix non properly initialized page headers.
In each map, the first page is reserved and the header pointing to it is
always supposed to be filled with 0. Since the header is used as a guard,
we always allocate the first header of the next map when setting up a new
map. We never did set the first header of the first map up, leading for it
to be allocate the first time we allocate a map.

But, when allocating a new map, we do p_realloc0(o + 1, n + 1) (old number
of page + 1, new number of page + 1) meaning the header 0 was never
properly initialized to 0. In practice it is always properly initialized
to 0 because it is one of the first allocations done in the life of the
process but *this is pure random*. We now properly allocate the first
header (this cost an extra heap allocation that will soon be reallocated,
but we really don't care), that way we are sure the first header is
*always* properly initialized.

Spotted thanks to the unit tests and improves consistency policy.

Change-Id: I86fe5186d154c2df39ccb0e43c920aee64eda661
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>