qkv: wah: improve invariant checking.
Does the check at the beginning and the end of each public function and
not in places where the wah can be in an inconsistent state because some
part of the operation have been applied while other remain pending.

Also add a check that the number of active (set to 1) bits is smaller than
the size of the bitmap.

Change-Id: I0b2e9bf27f971bf2ede7bb9ab5a8acd8ef7a7a24
Signed-off-by: Sebastien Boulle <sebastien.boulle@intersec.com>