license: introduce the encryptionKey
The encryptionKey will be generated by intersecid-sign from a given
plaintext key and the computed license signature.

This key will be used by products to initialize their SSL context to
encrypt things.

This patch add two functions, one to compute the obfuscated encryption key
from the real key and the computed signature. The other to resolve the
obfuscated key inside a product.

Note: these functions have been added to the ssl module to avoid
force openssl as a dependency of all products.

Refs: #9643
Change-Id: I527b8f7b58e5d5d6abf1cdb6c32d7a581737dc43