lstr: add a helper to map a file as a lstr_t.
First, we introduce a new mem flag type "MEM_MMAP" for memory issued from
a mmap. That flag has a value of 4 meaning we now need 3 bits instead of 2
to store the mem_pool flag in structures that support the new MEM_MMAP
flag (for now, only lstr_t needs that support).

Also add lstr_init_from_file() that map a file in a lstr_t and set the
correct flag.  That function opens the file, does the fstat and the mmap
and properly initialize the lstr_t. Since the flag MEM_MMAP is properly set,
lstr_wipe automatically calls munmap.

Change-Id: I965ebba6cef7c011aa2d7f05c0121d0b5682aaa4
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>