lstr: add lstr_persists function to force heap-allocation.
This function can be used to force a lstr_t to be heap-allocated in case
it is required. If the lstr_t is already heap-allocated and owns the
memory, it does nothing.

This function is not perfect since we currently have no way to distinguish
immediates and other global strings from not-owning copies of lstr_t (this
may require an improvement of the whose mem_pool stuff).

Change-Id: Ibcf1f8df13e99754b7f6b4b25a4770423e7389ff
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>