iop-json: handle negative double
Previously we considered leading - as the beginning of an integer
expression and we fed the operator to the constant folder (which only
supports integer).

Now we handle it as a number (so if it's a double we return it, if it's
an integer we feed it to the constant folder).

Closes: #10238
Change-Id: I6ab9e1f0c227e70b8f684db88c737a0815f55edb