errors: add e_assert() macros.
These macros aims at providing a simple interface to have assertions that
remain activated in production. These are not a remplacement for the
traditional assert(), it's just a helper to make writing correct error
handling easier and to improves the quality of the feedback for fatal
errors.

The usage is:
  e_assert(panic, blah && blih, "this should not happen");

(and the variant that won't dump a core: e_assert(fatal, ...), error,
warning, info, notice are also valid levels, but they won't stop the
program, so be careful, using them may be confusing).

If the condition is not verified, you'll get an output that looks like:

 blah-main.c:42:main: assertion failed: "blah && blih": this should not happen

In order to make some kind of checks easier, the following variants are
provided:
* e_assert_n() checks that the expression is not negative
* e_assert_neg() checks that the expression is negative
* e_assert_p() checks that a pointer is non-NULL
* e_assert_null() checks that a pointer is NULL

Change-Id: I7f0d7571dabd958c3a828376c1f20c80acec04f2
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>