i-channel: fix ic_is_ready()/ic_watch_activity()
ic_is_ready() returns false if the timer for the i-channel reconnection is
set.

Problem
~~~~~~~
Currently, the timer used for the reconnection is ichannel_t::timer. But
this timer is also used to send keep-alive packets (NOP) when
ic_watch_activity() has been called on the ichannel_t object.

Consequently, if we use ic_watch_activity(), ic_is_ready() may return
false whereas the connection is correctly established.

Solution
~~~~~~~~
Use a boolean to give the “queuable” state of the IChannel and fix the
ic_is_ready() condition.

Closes: #10613
Change-Id: Ic28060c228c009e3802c7a965905aeb0e4d2724b
Signed-off-by: Jonathan Squirawski <jonathan.squirawski@intersec.com>