mem: add support for alignment in the libc allocator.
This add the ability to add an alignment constraint on the memory
allocated on the heap. When no constraints is given, the allocation is
performed using malloc()/calloc(), when an alignment constraint is
provided the allocation is done using posix_memalign().

The alignment must be a power of two (this is actually a requirement of
posix_memalign()).

The tricky part of the patch is on the reallocation. realloc() has no
guarantee to preserve the alignment of the pointer (but specific
implementation may preserve it) and there's no aligned_realloc() call
defined. Thus , we first try to perform a reallocation using realloc(), if
that reallocation does not returns an aligned pointer, we do a brand new
allocation and copy the content of the realloced pointer in the new
allocation (that is costly but this may not happen too frequently).

Change-Id: I25d7e06b2081329ce5344068db0592261d586c33
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>