qps: use anonymous memory for non-ro maps.
New maps are allocated as anonymous maps instead of being backed by files.
The files get written to disk at snapshot time, using the exact same
format that was used previously (ensuring backward compatibility with
existing spools). When the snapshot is finished, the new file is mapped in
the exact place where the anonymous map previously was. As a consequence,
read-only data remains backed by the file system having less memory
pressure (since the rarely used data can be cheaply unloaded from the main
memory).

As a consequence, mutable memory maps are not backed by files anymore and
thus are not subject to page write-back anymore. This should reduce the
I/O load caused by the QPS in both QKV and QRRD (QPS now writes on disk
only during snapshot, only reads are possible the remaining of the time,
with enough RAM those reads are normally cached). Another consequence is
that the memory pressure of the QPS might slightly increase since more
maps are not backed by files anymore.

Also change so closes to the safer p_close() (that ensure the call is
retried if interrupted but is non-fatal in case of error).

Change-Id: Ic574c2e7f725c8c5aa33bacf644522df712e79a7
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>