qps: on-disk data are read-only.
Enforce the written data to be read-only. That way, we are sure the qps
cannot get corrupted by a running process (unless somebody explicitly
chmod the files).

The main consequence on the code is that we must ensure the header of each
map is read-write since it stores some runtime-dependent data (such as the
pointer of the QPS or an up-to-date value of the remaining used size),
this force us to map that specific page as anonymous memory and copy into
it the actual content of the disk (+ add runtime specific information).

Also remove some code on the load-meta that was due to the fact the header
of the page could be written back to disk. Since we now make an anonymous
map for that header, we're sure setting map->hdr.remaining won't affect
other processes and thus there's no need to have special code to handle
this.

Also, wraps munmap in the disk error detector to detect misuses of the
system call (minor).

Change-Id: I18cf111579b66ec080febdd173282f1cfb488616
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>