sort: implements optimized sort/uniq/seaches on 8 and 16 bits integers.
The implementation differs for sort and uniq on 8 bits. Instead of using a
dsort algorithm, we use a simple count of elements for sorting (since the
number of elements is limited, this is efficient in both CPU and memory)
and a bitfield of present elements for uniq.

Also add tests for all the data sizes.

Change-Id: I74cf58ed7ad87f497ba532b020efa690de529e98
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>