mem: t_* and p_* allocator use the natural alignment of the type.
Instead of using the minimal alignment requirement (sizeof(void *)), the
allocators will now use the natural alignment requirement of the allocated
structure. For most structures, this won't change anything since the
largest natural requirement is 64bits, put as soon as you added some
explicit alignment constraints on a structure, this is worth it since you
won't have to deal with the alignment manually.

Change-Id: I9f5d1497048a9a4fb26da036489e134a95d66bfe
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>