unix: move psinfo_get_tracer_pid back to unix-psinfo.c
We don't want to need all the dependencies of el.h whenever we want to use
psinfo_get_tracer_pid, thus, move the function back to the unix namespace
but keep the helper function inlined in order not to break its stealth
when used to avoid our process to be traced.

BTW, el-showflags.c is supposed to support inclusion at several places in
the case (see it's documentation at the top of the file) but it did
contain an exported function which is incompatible with multiple
inclusion.

Change-Id: I4654e9f86848053f8bc705cccf55a900f9eb5456
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>