z: don't fork when running within gdb.
It's really annoying to need to set gdb's follow-fork-mode each time I
need to run a debugger on a unit test. So, just have a degraded mode when
running within gdb (or strace or whatever process that want to ptrace()
our unit tests) so that is just easier and straightforward to debug it.

As a consequence z is less robust when running within a tracer since a
crash in a test will crash everything but well... crash are there to be
debugged and if you want to inspect a particular bug just run only the bug
that interests you to avoid useless noise.

Change-Id: I0100f875c23b3b240aa0c9df96d1f5a194559c46
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>