core-mem-stack: fix regression in valgrind instrumentation.
In case of a growing realloc with a copy, we marked the some bytes outside
the old chunk as no access (starting at the beginning of the old chunk +
the new size). As a consequence, the place marked zone could belong to
another allocation leading to invalid bad access notification under
valgrind.

Mark correctly the padding of the *new* chunk as no access.

This is a regression from af8d76aa19d.

Change-Id: I7f78be03818776196f1895d78434f93715df1f9c
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>