wah: simplify the handling of the first run.
Previously, the first run was stored inlined in the structure leading to
some hacks in order to detect if we needed to read that first run or the
following ones (stored in a qv_t).

We lose a bit of performance since we need an extra allocation each time
we use a wah (even trivial ones), but this will be fixed in a future
commit (it requires a patch in QKV before). The lose is actually minor
since currently, all wah_t users (QKV) use pools of wah_t to avoid
allocations.

Change-Id: I0f91303788323eb591aa720418e5c76c73bee96b
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>