asn1: adapt CHOICE registering helper to make it compatible with IOP unions
Creates the macro __ASN1_IOP_CHOICE_DESC_BEGIN(). The two underscores
indicating that the macro is unsafe are here because there are prerequisites
on the IOP union tagging: tagging must start with value 1 and be strictly
incremental.

This patch allows the use of other integer types than enum for the union
selector, so some casts have been adapted.

-------------------------------------------
Examples:
-------------------------------------------

/* Bad */
union Toto {
   int a;
   string b;
   string c;
};

/* Good */
union Toto {
1: int a;
2: string b;
3: string c;
};

/* Good */
union Toto {
1: int a;
   string b;
   string c;
};

/* Bad */
union Toto {
1: int a;
2: string b;
4: string c;
};

-------------------------------------------

Why?
~~~~
The motivation to increase interoperability between Intersec ASN.1 packer
and IOP are to benefit of the IOP helpers and utilities and to avoid
hand writing of C-ASN.1/IOP converters in anticipation of the fact we could
want to send C-ASN.1 structure through IOP channels at midterm.

The perfect way to ensure this interoperability would be to develop a
BER/PER module for IOPs, unfortunately, the IOP syntax is too weak for now
to implement PDUs for MAP or TCAP.

Change-Id: I19d08ca739a6e085c03b7a2d73a0636391eb4e58
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>