mem: reduce overhead of *_new_extra_field implementation.
The previous implementation added the size of the field to the size of the
structure. This can have an overhead if the field is not aligned on the
bottom of the structure:

  struct a {
     uint64_t a;
     uint32_t b;
     byte     c[];
  }

In that case offsetof(struct a, c) is not sizeof(struct a) because c is
aligned on 1 byte and put directly after the 4 bytes of b while struct a
is aligned on 8 bytes. As a consequence, p_new_extra_field(struct a, c 10)
allocated 4 useless bytes since it allocated
  sizeof(struct a) + sizeof(*c) * 10

The fix is to use the offset of the field as the base for the allocation
and add the size of the extra fields to that base. In order to ensure we
can p_clear() the structure, we also ensure we never allocate less than
the size of the structure.

Change-Id: If595607a9b7f76916bc63d8f37fda756c877653b
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>