qvector: add qv_for_each_ptr and qv_for_each_entry.
Those two macros does the same as qv_for_each_pos, but instead of
enumerating the positions, they enumerated the values contained in the
qvector (either as pointers or as plain types).

Consider these macros as experimental for now in order to check if they
can help improve our code (or on the contrary trigger unexpected bugs).

Change-Id: Ie287c03b5459f35a9624ddf0b634bb504c8b3e7a
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>