qvector: fix compilation of qv_for_each_entry on old gccs.
Old version of gcc are not able to detect that 'e' is not used when
(vec)->len is 0, as a consequence, they emit a warning about 'e' being
used uninitialized in the loop.

To fix this, ensure e is always properly initialized, by fetching the
value from the vector whenever possible (possibly avoiding a call to
memset), and calling p_clear() in the other cases.

Change-Id: I4a48f98ee049e2173bdbf710cc05c39bc5ec819a
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>