iop-rpc-channel: remove dangerous cast in ic_msg_p()
The macro ic_msg_p() helps using private data when calling IOP RPCs.
Allocating the right size for the private data using the type given as a
first argument.

The macro, in its first and current implementation, includes a cast of the
private data parameter, getting rid of the type check, this is dangerous
because it can lead to errors like the one on patch set three of change
If14c489ad5f7159b40a4ff03dba165731e014be2:

* The developer had to save a slot of a pending IOP RPC
* He wrote ic_msg_p(uint64_t, slot)
* He should have written ic_msg_p(uint64_t, &slot)
* His slot has been casted from uint64_t to uint64_t * and indirected,
  leading to a potential/probable segmentation fault and the loss of the
  real slot value

Without the cast, the compiler would have detected a assignation of a
pointer of type uint64_t * from an incompatible type uint64_t.

Change-Id: I44aa4d96342aa1744b31959d043c2c673a451ca9
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>