container-qhash: implement __qm_reserve[_h] helpers + tests
What?
~~~~~
Same thing as __qm_put[_h] without the value argument.

Why?
~~~~
In most cases, the macro __qm_put[_h]() is called with an nul value, the
final value being built and set in case there is no collision.

This is no big deal for pointed values, because we just have to use NULL:

  pos = __qm_put(foo, &foo_h, key, NULL, 0);
  /* Do stuff */

Though the value argument is pointless.

But when the value is a structure instance, the user has to initialize a
useless structure:

  foo_t tmp;

  /* Avoid invalid read */
  p_clear(&tmp, 1);

  pos = __qm_put(foo, &foo_h, key, tmp, 0);
  /* Do stuff */

With __qm_reserve():

  pos = __qm_reserve(foo, &foo_h, key, 0);
  /* Do stuff */

Change-Id: I9a9616fde729a1f0dcfd18d9c987137eec6b4e8d
Signed-off-by: Sébastien Thouément <sebastien.thouement@intersec.com>