unix: add fd_get_path()
That function get the path of a file or directory pointed by a fd. It is a
bit expensive since it checks its return value (a file could have been
replaced by another one). It is not supposed to be heavily used.

This function will help implementing callat() variants of functions when
the native system call is missing.

Change-Id: I41299a651c5a0488497c086ca4d1a6e9ce2975c4
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>