unix: implements mkdirat_p() and rmdirat_r().
Those functions do the same as the regular variant but interpret the path
as relative to the given file descriptor.

The mkdirat_p() implementation uses the native mkdirat() system call while
the rmdirat_r() implementation has to use the new fd_get_path() function
in order to provide arguments compatible with the fts mode.

Change-Id: I78fe9e8dfd36426b364d8d432d6485800039ec9d
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>