thr-job: fix main thread not being waken up when queuing events.
When we push an event in the main queue while on another thread, we call
thr_wakeup_thr0() that send a SIGCHLD to the main thread. That signal aims
at interrupting the epoll_wait() and restarts an event loop. Since the
consumption of the main queue is done in a el_before() trigger, the
SIGCHLD should trigger an (nearly) immediate consumption of the main
queue.

However, a signal cannot interrupt epoll_wait() if no sighandler is
associated to that signal. Thus if no-one is waiting for a SIGCHLD in the
process, the SIGCHLD is just ignored and does not wake the event loop up.
As a consequence, the job posted won't get executed before the current
loop has ended and a new one has been started. This can take up to 59s
(default timeout of the event loop) if no timer is registered and no
activity occurs on the registered fds.

To fix this, thr_job now always ensure there's a handler for SIGCHLD.
Since the event loop supports several el_t per signal, since does not
affect the rest of the world (unless some code explicitly ignored
SIGCHLD).

Change-Id: I79968359693f50fe71d29ca26aac7bbb6a2d9af3
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>