core-macros: introduces THROW_*_IF() macros.
And deprecates RETURN_*_IF() macros at the same time. The new naming is
much more explicit about the expected use case: error emission in case a
condition is not verified.

The RETURN_*_IF() macros are kept for now in order to keep code
compatibility with existing sources, but they will be removed in
lib-common/master. New code *MUST NOT* use the RETURN_*_IF() macros
anymore.

Change-Id: I1feee5b7c0fb2a89388717575db9aec407e5843e
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>