iop: simplify the iop_equals() function and add unit tests
These tests need an iopc >= 2.3.3 to work correctly since the
iop_equals() function was broken because it relied on non-given meta-data
about scalar field sizes.

The iopc 2.3.3 now gives the scalar sizes in the field description, so we
can add test to prove it works.

We also take advantage of the new iopc feature to simplify and optimize
the optional fields handling (which are now comparable with a single
memcmp).

Closes: #12256
Change-Id: I2a56ba0f92c194a340f24176d3d7b29b8e7e7bae