have helpers to lock/unlock a directory
These functions provide a way to lock a directory and avoid concurrent
accesses. lockdir() tries to create a .lock file and must be called before
managing the directory. This function returns an error if the directory is
already handled by another process. unlockdir() must be called when
releasing the directory.

See function documentation for further information.

This mechanism is already used in qdb and nearly in MCG. Implementing it
in lib-common will avoid code duplication.

Change-Id: Ie9cd61bceceaf2a86333cc40768b9b8b90edc3f7
Signed-off-by: Vianney Bajart <vianney.bajart@intersec.com>