z: improves behavior under gdb and valgrind.
In order to make it easier to trace the tests with valgrind, don't fork
(unless asked otherwise) when running on valgrind.

The same way, force use of abort in case of assertion failure when the
tests get traced (and as a consequence when they're run on gdb).

Also add the proper command line flags in order to be able to tweak the
default behavior.

Change-Id: I322e5ad86d7076f288821d821294ad9d1dc3aa35
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>