wah: add wah_add1_at in order to set a specific bit.
This function should only be used for bit offsets greater than the current
size of the wah since wah are not supposed to handle random accesses
(which mean a full copy of the wah).

Keep the inefficient copy mode in production mode in order to maintain
sain bitmaps (at the expense of the performances).

Change-Id: I1a3bd3d39701221ccc827f2d7bc44fd3bcb6c450
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>