wah: add wah_and_not() and wah_not_and() operators.
These operator takes two wah A and B and builds A & ~B (resp. ~A & B),
which is useful to perform boolean expression evaluation.

As a consequence, the ability to enumerate the reverse of the reverse of
the wah has been moved from the bit enumerator to the word enumerator,
breaking the public API.

Change-Id: I77deca5573d06521dbfb979f24237841fbdd9f28
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>