mem-ring: fix segmentation fault when merging frame
The ring_reset_to_prevframe() function used to do:

    fprev->rp &= ~FRAME_IS_FREE;

Which is obviously wrong with FRAME_IS_FREE define to 1U… Fortunately,
with most GCC FRAME_IS_FREE is taken as a 64bits value. But sometimes it's
compiled with a true 32bits value and when you get a pointer using more
than 32bits here… you crash.

Change-Id: Ic3cc1c716a9045bfc531704cc03fe06b3ccb7387