wah: fix "fast" mode in binary op.
We supposed the number number of bits that remained to be added were the
difference between the two wahs, while it is the difference between the
largest wah and the number of the bit current added. The previous
approach could not work because run larger than the remain bits from a wah
can be written.

Change-Id: I22eee663a61495e417823a6f85e45871a2044c4b
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>