Build: generate dependencies files for .blk.
This will avoid blk files not being generated when some structures it
depends on get modified (this issue happened on LMS and caused a cascade
of build failures).

The downside here is that we cannot use the call to the rewriter to
generate those files (the -M* options are not available when -cc1 is
passed to clang, which is required for -rewrite-blocks...). As a
consequence, we now have two calls to clang in the block rewriting code.

Change-Id: Iefc7ec5f07c3726196fc081bf71044174c4705b9
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>