lib-common: add qv_init_static macro
The path aim is to allow to init a qv with a not owned memory.
Typical used could be to init a qv with data which already exist.

As for the conversion of an IOP array into a qv_t() when an
API requires a qv_t(),
or in APIs that take a (pointer, count) and need to forward it
to a qv-based API,
or simply to make qv initialization easier:
 uint32_t data[] = { 1, 2, 3, 4 };
 qv_t(u32) vec;
 qv_init_static(u32, &vec, data, countof(data));

Be aware that if the qv is extend the data will be dump on
the libc pool, so don't forget to wipe it and maybe this is
not the init you want.

Change-Id: Iaaea0d5fc40f064b5f625723884f398777f89334
Signed-off-by: Sebastien Boulle <sebastien.boulle@intersec.com>