lib-common: wah, add missing map reset in acquire
Before being returned, an old map kept in the pool
should be reset. This way the user get a ready to
use map.

The reset could be done either in the acquire or in
the release.
However maps could be big, so to avoid memory overuse
they will be reset at release time.

Change-Id: I10294ce548dfaeffd9dadcb19f0c044237fa9f7b
Signed-off-by: Sebastien Boulle <sebastien.boulle@intersec.com>