iop[json]: fix potential t-stack issue in iop_jpack()
The iop-jpack() function use a local stack frame to dump the IOP data
fields. But the IOP should never use a local stack frame to give the
ability to its user to use the t_pool() to pack or unpack every kind of
IOP structures.

Here if you used a t-allocated buffer to pack your JSon you loosed…

Change-Id: I1847af56e030674dba9d463f42d6217612e3c2ce