str-buf-gsm: fix sb_conv_to_gsm_isok
This function assumed that the first 128 utf8 char (on 7 bits) could be
mapped to gsm7, which is wrong because most characters below 0x20 are
control char which cannot be translated. However these char are unlikely
to happen, and I was worried about adding one function call in all
cases for every char. I avoided it by inlining win1252_to_gsm7

Note: the function is now closer to sb_conv_to_gsm

Also add plan as argument, it will not break anything since this
function is not used anywhere

Change-Id: Ic4aad3ea8dc38c58f5ef4c23b518d39f51d6affe