iop: Add a "flags" attribute to iop_rpc_t
This is done without modifying the ABI. The initial size of iop_rpc_t
is 48 due to struct alignements. The bool field is of size 1 and the
lstr_t field of size 16. After this patch the iop_rpc_t remains of
size 48 (at least in 64 bits), thus the ABI should not be modified.
This is done in the same spirit of the iop_struct_t flags.

Refs: #11378
Change-Id: I72e383677afb33eddd70060735122c4f9a713cdf
Progress: 60%
Signed-off-by: Nicolas Barbey <nicolas.barbey@intersec.com>